\name{xdotool-package}
\alias{xdotool-package}
\alias{xdotool}
\docType{package}
\title{
\packageTitle{xdotool}
}
\description{
\packageDescription{xdotool}
}
\details{
  An R interface to the \code{xdotool} command (on Linux).
}
\author{
\packageAuthor{xdotool}

Maintainer: \packageMaintainer{xdotool}
}
\references{
  \url{http://www.semicomplete.com/projects/xdotool/}
}
\keyword{ package }
\examples{
mousemove(200, 200)
keystroke("Ctrl")
}
